refactor login page and form components for improved structure and styling
